+++ uniffi_bindgen/src/bindings/python/templates/CallbackInterfaceImpl.py
@@ -1,91 +1,62 @@
 {% if self.include_once_check("CallbackInterfaceRuntime.py") %}{% include "CallbackInterfaceRuntime.py" %}{% endif %}
+{%- let trait_impl=format!("UniffiTraitImpl{}", name) %}
 
-# Declaration and _UniffiConverters for {{ type_name }} Callback Interface
+# Put all the bits inside a class to keep the top-level namespace clean
+class {{ trait_impl }}:
+    # For each method, generate a callback function to pass to Rust
+    {%- for (ffi_callback, meth) in vtable_methods.iter() %}
 
-def {{ callback_handler_class }}(handle, method, args_data, args_len, buf_ptr):
-    {% for meth in methods.iter() -%}
-    {% let method_name = format!("invoke_{}", meth.name())|fn_name %}
-    def {{ method_name }}(python_callback, args_stream, buf_ptr):
-        {#- Unpacking args from the _UniffiRustBuffer #}
-        def makeCall():
-            {#- Calling the concrete callback object #}
-            {%- if meth.arguments().len() != 0 -%}
-            return python_callback.{{ meth.name()|fn_name }}(
-                {% for arg in meth.arguments() -%}
-                {{ arg|read_fn }}(args_stream)
-                {%- if !loop.last %}, {% endif %}
-                {% endfor -%}
-            )
-            {%- else %}
-            return python_callback.{{ meth.name()|fn_name }}()
+    @{{ ffi_callback.name()|ffi_callback_name }}
+    def {{ meth.name()|fn_name }}(
+            {%- for arg in ffi_callback.arguments() %}
+            {{ arg.name()|var_name }},
+            {%- endfor -%}
+            {%- if ffi_callback.has_rust_call_status_arg() %}
+            uniffi_call_status_ptr,
             {%- endif %}
-
-        def makeCallAndHandleReturn():
-            {%- match meth.return_type() %}
-            {%- when Some(return_type) %}
-            rval = makeCall()
-            with _UniffiRustBuffer.alloc_with_builder() as builder:
-                {{ return_type|write_fn }}(rval, builder)
-                buf_ptr[0] = builder.finalize()
-            {%- when None %}
-            makeCall()
-            {%- endmatch %}
-            return _UNIFFI_CALLBACK_SUCCESS
+        ):
+        uniffi_obj = {{ ffi_converter_name }}._handle_map.get(uniffi_handle)
+        def make_call():
+            args = ({% for arg in meth.arguments() %}{{ arg|lift_fn }}({{ arg.name()|var_name }}), {% endfor %})
+            method = uniffi_obj.{{ meth.name()|fn_name }}
+            return method(*args)
+        {%- match meth.return_type() %}
+        {%- when Some(return_type) %}
+        def write_return_value(v):
+            uniffi_out_return[0] = {{ return_type|lower_fn }}(v)
+        {%- when None %}
+        write_return_value = lambda v: None
+        {%- endmatch %}
 
         {%- match meth.throws_type() %}
         {%- when None %}
-        return makeCallAndHandleReturn()
-        {%- when Some(err) %}
-        try:
-            return makeCallAndHandleReturn()
-        except {{ err|type_name }} as e:
-            # Catch errors declared in the UDL file
-            with _UniffiRustBuffer.alloc_with_builder() as builder:
-                {{ err|write_fn }}(e, builder)
-                buf_ptr[0] = builder.finalize()
-            return _UNIFFI_CALLBACK_ERROR
+        _uniffi_trait_interface_call(
+                uniffi_call_status_ptr.contents,
+                make_call,
+                write_return_value,
+        )
+        {%- when Some(error) %}
+        _uniffi_trait_interface_call_with_error(
+                uniffi_call_status_ptr.contents,
+                make_call,
+                write_return_value,
+                {{ error|type_name }},
+                {{ error|lower_fn }},
+        )
         {%- endmatch %}
+    {%- endfor %}
 
-    {% endfor %}
+    @{{ "CallbackInterfaceFree"|ffi_callback_name }}
+    def uniffi_free(uniffi_handle):
+        {{ ffi_converter_name }}._handle_map.remove(uniffi_handle)
 
-    cb = {{ ffi_converter_name }}._handle_map.get(handle)
-
-    if method == IDX_CALLBACK_FREE:
-        {{ ffi_converter_name }}._handle_map.remove(handle)
-
-        # Successful return
-        # See docs of ForeignCallback in `uniffi_core/src/ffi/foreigncallbacks.rs`
-        return _UNIFFI_CALLBACK_SUCCESS
-
-    {% for meth in methods.iter() -%}
-    {% let method_name = format!("invoke_{}", meth.name())|fn_name -%}
-    if method == {{ loop.index }}:
-        # Call the method and handle any errors
-        # See docs of ForeignCallback in `uniffi_core/src/ffi/foreigncallbacks.rs` for details
-        try:
-            return {{ method_name }}(cb, _UniffiRustBufferStream(args_data, args_len), buf_ptr)
-        except BaseException as e:
-            # Catch unexpected errors
-            try:
-                # Try to serialize the exception into a String
-                buf_ptr[0] = {{ Type::String.borrow()|lower_fn }}(repr(e))
-            except:
-                # If that fails, just give up
-                pass
-            return _UNIFFI_CALLBACK_UNEXPECTED_ERROR
-    {% endfor %}
-
-    # This should never happen, because an out of bounds method index won't
-    # ever be used. Once we can catch errors, we should return an InternalException.
-    # https://github.com/mozilla/uniffi-rs/issues/351
-
-    # An unexpected error happened.
-    # See docs of ForeignCallback in `uniffi_core/src/ffi/foreigncallbacks.rs`
-    return _UNIFFI_CALLBACK_UNEXPECTED_ERROR
-
-# We need to keep this function reference alive:
-# if they get GC'd while in use then UniFFI internals could attempt to call a function
-# that is in freed memory.
-# That would be...uh...bad. Yeah, that's the word. Bad.
-{{ callback_handler_obj }} = _UNIFFI_FOREIGN_CALLBACK_T({{ callback_handler_class }})
-_UniffiLib.{{ ffi_init_callback.name() }}({{ callback_handler_obj }})
+    # Generate the FFI VTable.  This has a field for each callback interface method.
+    uniffi_vtable = {{ vtable|ffi_type_name }}(
+        {%- for (_, meth) in vtable_methods.iter() %}
+        {{ meth.name()|fn_name }},
+        {%- endfor %}
+        uniffi_free
+    )
+    # Send Rust a pointer to the VTable.  Note: this means we need to keep the struct alive forever,
+    # or else bad things will happen when Rust tries to access it.
+    _UniffiLib.{{ ffi_init_callback.name() }}(ctypes.byref(uniffi_vtable))