+++ uniffi_bindgen/src/bindings/python/templates/CallbackInterfaceTemplate.py
@@ -1,10 +1,10 @@
 {%- let cbi = ci|get_callback_interface_definition(name) %}
-{%- let callback_handler_class = format!("UniffiCallbackInterface{}", name) %}
-{%- let callback_handler_obj = format!("uniffiCallbackInterface{}", name) %}
 {%- let ffi_init_callback = cbi.ffi_init_callback() %}
 {%- let protocol_name = type_name.clone() %}
-{%- let methods = cbi.methods() %}
 {%- let protocol_docstring = cbi.docstring() %}
+{%- let vtable = cbi.vtable() %}
+{%- let methods = cbi.methods() %}
+{%- let vtable_methods = cbi.vtable_methods() %}
 
 {% include "Protocol.py" %}
 {% include "CallbackInterfaceImpl.py" %}