+++ CHANGELOG.md
@@ -12,6 +12,11 @@
  
 - Fixed a memory leak in callback interface handling.
 
+### ⚠️ Breaking Changes for external bindings authors ⚠️
+
+- The callback interface code was reworked to use vtables rather than a single callback method.
+  See https://github.com/mozilla/uniffi-rs/pull/1818 for details and how the other bindings were updated.
+
 ## v0.26.0 (backend crates: v0.26.0) - (_2024-01-23_)
 
 ### What's changed?